worker/requirements.txt
torch==2.10.0
openai-whisper
ffmpeg-python
